+++ CHANGELOG.rst
@@ -35,6 +35,12 @@ Version 3.5.1.post0
 ====================
 Unreleased
 
+Added
+-----
+* Run a subset of jobs by adding their index number(s) as command line arguments.  For example, run ``webchanges 2
+  3`` to only run jobs #2 and #3 of your job list.  Run ``webchanges --list`` to find the job numbers. Suggested by
+  `dbro <https://github.com/dbro>`__ upstream `here <https://github.com/thp/urlwatch/pull/641>`__.
+
 Fixed
 -----
 * readthedocs.io fails to build autodoc API documentation